requirements.txt
flask==3.1.0
openpyxl==3.1.5
gunicorn==23.0.0
psycopg2-binary==2.9.12
authlib==1.6.0
requests==2.32.3
